requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
setuptools==70.0.0
SQLAlchemy==2.0.30
svgwrite==1.4.3
typing_extensions==4.12.1
Werkzeug==3.0.3
WTForms==3.1.2
